pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"

[tool.poetry.dependencies]
python = ">=3.11,<4"
email-validator = ">=2.2.0"
flask-limiter = {version = ">=3.10.1", extras = ["redis"]}
flask = ">=3.1.0"
flask-sqlalchemy = ">=3.1.1"
gunicorn = ">=23.0.0"
psycopg2-binary = ">=2.9.10"
flask-wtf = ">=1.2.2"
sqlalchemy = ">=2.0.38"
flask-login = ">=0.6.3"
redis = ">=5.2.1"
flask-apispec = ">=0.11.4"
marshmallow = ">=3.26.1"
apispec = ">=6.8.1"
webargs = ">=8.6.0"
werkzeug = ">=3.1.3"
wtforms = ">=3.2.1"
uvicorn = ">=0.34.0"
python-multipart = ">=0.0.20"
pydantic = ">=2.10.6"
python-jose = {version = ">=3.4.0", extras = ["cryptography"]}
passlib = {version = ">=1.7.4", extras = ["bcrypt"]}
fastapi = ">=0.115.8"
oauthlib = ">=3.2.2"
routes = ">=2.5.1"
twilio = ">=9.4.6"
starlette = ">=0.45.3"
python-dotenv = ">=1.0.1"
requests = ">=2.32.3"

[tool.poetry]
package-mode = false
